feat: support graphic connector
The rendering material is automatically generated. You don't need to pre-generate the material.
The generated materials are cached and properly batched.

Close #75, close #76, close #80

BREAKING CHANGE: The name of the custom SoftMaskable shader must be changed. For more information, see the ‘Support soft masks with your custom shaders’ section of the README.